Refactor: navigate 프로필 수정으로 가는 경로 수정, 초기 UserInfoAPI이름 수정 (#15)
